+++ backend/src/config/redisconfig.ts
@@ -2,8 +2,8 @@
 import Redis from "ioredis";
 
 const redis = new Redis({
-  // host: "localhost",
-  host: "redis",
+  host: "localhost",
+  // host: "redis",
   port: 6379,
 });
