+++ backend/src/routes/userRoutes.ts
@@ -12,8 +12,8 @@ const userRouter = Router();
 
 userRouter.get("/", verifyAdminAndGuard, getUsers);
 userRouter.post("/", verifyAdmin, createUser);
+// userRouter.post("/", createUser);
 userRouter.get("/:userId", verifyAdminAndGuard, getUserById);
 userRouter.patch("/:userId", updateUser);
 userRouter.delete("/:userId", verifyAdmin, deleteUser);
-
 export default userRouter;
